taxeen/backend/requirements.txt
# Taxeen Backend Requirements
# FastAPI + SQLAlchemy + PostgreSQL + Redis + Celery

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
passlib[argon2]==1.7.4
pyjwt==2.8.0
cryptography==42.0.0

# PDF Processing
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.2.0
opencv-python-headless==4.9.0.80

# Task Queue
celery[redis]==5.3.6
redis==5.0.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
